sale_automatic_workflow: Do not post invoices set back to draft
When sale automatic workflow is set to create and post invoices,
a posted invoice that is set back to draft by the user will be
posted again automatically on the next run of the sale automatic
workflow.

This change in the filter avoids to post the invoice automatically
and allows the user to do its changes before posting it manually.